edoreczenia/requirements.txt
annotated-types==0.7.0
argcomplete==3.6.2
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
datamodel-code-generator==0.33.0
dnspython==2.8.0
email-validator==2.3.0
genson==1.3.0
idna==3.10
inflect==7.5.0
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
more-itertools==10.8.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
pytokens==0.1.10
PyYAML==6.0.2
requests==2.32.5
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
